requirements.txt
numpy==1.19.5
opencv_python==4.7.0.72
PyYAML==6.0
tensorflow==2.6.0
timm==0.6.13
torch==2.0.0
tqdm==4.64.0
protobuf==3.20.*